+++ src/main/scala/com/jcdecaux/datacorp/spark/workflow/Flow.scala
@@ -1,6 +1,8 @@
 package com.jcdecaux.datacorp.spark.workflow
 
-private[workflow] case class Flow(payload: String, from: Node, to: Node, stage: Int) {
+import scala.reflect.runtime
+
+private[workflow] case class Flow(payload: runtime.universe.Type, from: Node, to: Node, stage: Int) {
   def describe(): Unit = {
     println("Flow")
     println(s"Stage     : $stage")